pyproject.toml
[project]
name = "projects-aguest-me"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "beanie>=1.29.0",
    "fastapi[standard]>=0.115.11",
    "gunicorn>=23.0.0",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
    "uvicorn-worker>=0.3.0",
]
